Size no longer needs to be a multiple of number of tasks.
The MPI support was enhanced so that the work can be distributed over a number of tasks that doesn't divide evenly into the array lengths.  There are a few tests that are failing, which still need to be tracked down.  Note, these are all for cases where the transform length is even, which is a situation that might be better avoided, altogether, as discussed in the documentation.