Added a box_size parameter to SpectralGrid.
The periodic box can now be of arbitrary size, rather than needing to be 2 \pi in each direction.  Note, with the lifting of this constraint, the wavenumber coefficeints for a Taylor-Green vortex are no longer necessarily integers.  They should be in a certain ratio to the box size, but this is no longer enfored.